Viewport, player's design, full screen
For this update, I changed the player's sprite to a new 32x32 one. In the future I will add animations.

I also changed the resolution to 320x240, and added a viewport which follows Player, centering her on x but y being a negative number so Player appears at the bottom of the screen. The code isn't perfect, but it' will do for now.

The last thing I added is that when F4 is pressed, it becomes fullscreen without losing image quality.